Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > just wondering but why would you even build these tests?
> > 
> > Strange to me too. I don't even have them in my mk files. So what happened is, I updated the proprietary blobs to A11 and tried building. It was then that I encountered this issue.
> 
> How is that even possible? I guess that no one here is opposed to the change itself, but it's just strange.

Do we have any reference devices close to msm8953/sdm632 that has got the updated list of 'proprietary-files.txt' so that I can alter accordingly? Maybe I might have added some library which is not required or something like that. Or, if it doesn't affect anything else then can we proceed with this change?

Patch-set: 1